Extract `ShareLinks` sub-component for reuse
