+++ src/styles/sidebar/components/share-links.scss
@@ -1,0 +1,10 @@
+@use '../../mixins/links';
+
+.share-links {
+  @include links.footer-links;
+
+  &__icon {
+    width: 24px;
+    height: 24px;
+  }
+}
