+++ src/styles/sidebar/sidebar.scss
@@ -45,6 +45,7 @@ $base-line-height: 20px;
 @import './components/selection-tabs';
 @import './components/share-annotations-panel';
 @import './components/search-input';
+@import './components/share-links';
 @import './components/sidebar-panel';
 @import './components/svg-icon';
 @import './components/spinner';
